chore: create models and tokens folders to move entities into appropriate folders
