+++ projects/angular8-yandex-maps/src/lib/components/ya-control/ya-control.directive.spec.ts
@@ -4,7 +4,7 @@ import { BehaviorSubject } from 'rxjs';
 
 import { YaControlDirective, YaControlType } from './ya-control.directive';
 import { YaMapComponent } from '../ya-map/ya-map.component';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import {
   createMapSpy,
   createRoutePanelConstructorSpy,
