+++ projects/angular8-yandex-maps/src/lib/components/ya-geoobject/ya-geoobject.directive.spec.ts
@@ -4,7 +4,7 @@ import { BehaviorSubject } from 'rxjs';
 
 import { YaGeoObjectDirective } from './ya-geoobject.directive';
 import { YaMapComponent } from '../ya-map/ya-map.component';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import {
   createGeoObjectConstructorSpy,
   createGeoObjectSpy,
