+++ projects/angular8-yandex-maps/src/lib/components/ya-multiroute/ya-multiroute.directive.spec.ts
@@ -4,7 +4,7 @@ import { BehaviorSubject } from 'rxjs';
 
 import { YaMapComponent } from '../ya-map/ya-map.component';
 import { YaMultirouteDirective } from './ya-multiroute.directive';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import {
   createMapSpy,
   createMultirouteConstructorSpy,
