+++ projects/angular8-yandex-maps/src/lib/components/ya-panorama/ya-panorama.directive.spec.ts
@@ -4,7 +4,7 @@ import { BehaviorSubject, Observable } from 'rxjs';
 
 import { YaMapComponent } from '../ya-map/ya-map.component';
 import { YaPanoramaDirective } from './ya-panorama.directive';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import {
   createMapSpy,
   createPlayerConstructorSpy,
