+++ projects/angular8-yandex-maps/src/lib/components/ya-placemark/ya-placemark.directive.spec.ts
@@ -4,7 +4,7 @@ import { BehaviorSubject } from 'rxjs';
 
 import { YaMapComponent } from '../ya-map/ya-map.component';
 import { YaPlacemarkDirective } from './ya-placemark.directive';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 import {
   createMapSpy,
   createPlacemarkConstructorSpy,
