+++ projects/angular8-yandex-maps/src/lib/components/ya-placemark/ya-placemark.directive.ts
@@ -11,10 +11,10 @@ import {
 } from '@angular/core';
 import { Observable, Subscription } from 'rxjs';
 
-import { EventManager } from '../../event-manager';
-import { YaEvent } from '../../typings/ya-event';
+import { EventManager } from '../../event-manager/event-manager';
+import { YaEvent } from '../../models/ya-event';
 import { YaMapComponent } from '../ya-map/ya-map.component';
-import { YaReadyEvent } from '../../typings/ya-ready-event';
+import { YaReadyEvent } from '../../models/ya-ready-event';
 
 /**
  * The `ya-placemark` directive wraps `ymaps.Placemark` class from the Yandex.Maps API.
