+++ projects/angular8-yandex-maps/src/lib/event-manager/event-manager.spec.ts
