+++ projects/angular8-yandex-maps/src/lib/models/ya-event.ts
