+++ projects/angular8-yandex-maps/src/lib/models/ya-ready-event.ts
