+++ projects/angular8-yandex-maps/src/lib/services/ya-api-loader/ya-api-loader.service.ts
@@ -1,46 +1,10 @@
 import { DOCUMENT } from '@angular/common';
-import { Inject, Injectable, InjectionToken, Optional } from '@angular/core';
+import { Inject, Injectable, Optional } from '@angular/core';
 import { from, fromEvent, merge, Observable, throwError } from 'rxjs';
 import { map, switchMap, take } from 'rxjs/operators';
 
-/**
- * Injection token to specify configuration.
- */
-export const YA_CONFIG = new InjectionToken<YaConfig>('YA_CONFIG');
-
-/**
- * API loading parameters.
- */
-export interface YaConfig {
-  /**
-   * API key. You can get a key in the developer's dashboard.
-   */
-  apikey?: string;
-  /**
-   * Locale.
-   */
-  lang?: 'ru_RU' | 'en_US' | 'en_RU' | 'ru_UA' | 'uk_UA' | 'tr_TR';
-  /**
-   * The order for setting geographical coordinates in API functions that accept longitude-latitude input.
-   */
-  coordorder?: 'latlong' | 'longlat';
-  /**
-   * List of modules to load.
-   */
-  load?: string;
-  /**
-   * API loading mode.
-   */
-  mode?: 'release' | 'debug';
-  /**
-   * Use commercial version of the API.
-   */
-  enterprise?: boolean;
-  /**
-   * Version number of the API.
-   */
-  version?: string;
-}
+import { YA_CONFIG } from '../../tokens/ya-config';
+import { YaConfig } from '../../models/ya-config';
 
 /**
  * The `YaApiLoader` service handles loading of Yandex.Maps API.
